backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
langchain-core
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
